requirements.txt
aiohttp==3.6.2
apiai==1.2.3
async-timeout==3.0.1
asyncio==3.4.3
discord.py==1.4.1
qrcode==6.1
requests==2.20.0
httplib2==0.18.1
pytz>=2013b
